- updated hadoop image (one)

+++ contrib/opennebula/hadoop-iso/config.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
 
+#  -m http://cdn.debian.de/debian/ \
+#  --bootloader syslinux --syslinux-menu false \
+
+
+
 lh config \
-  --distribution stable --memtest none --binary-indices false \
-  --syslinux-timeout 1 --archive-areas "main contrib non-free" \
+  --syslinux-timeout 1 --binary-indices false \
+  --archive-areas "main contrib non-free" \
+  --distribution stable --memtest none \
   --iso-application "Hadoop Live" \
   --iso-publisher "Contrail Project; http://contrail-project.eu" \
   --cache-indices true \
@@ -17,16 +23,20 @@ cp cust/cloudera.list config/chroot_sources/cloudera.chroot
 cp cust/cloudera.list config/chroot_sources/cloudera.binary
 cp cust/cloudera.gpg  config/chroot_sources/cloudera.chroot.gpg
 cp cust/cloudera.gpg  config/chroot_sources/cloudera.binary.gpg
+cp cust/xtreemfs.list config/chroot_sources/xtreemfs.chroot
+cp cust/xtreemfs.list config/chroot_sources/xtreemfs.binary
+cp cust/xtreemfs.gpg  config/chroot_sources/xtreemfs.chroot.gpg
+cp cust/xtreemfs.gpg  config/chroot_sources/xtreemfs.binary.gpg
 
 # xtreemfs client
-test -f config/chroot_local-packages/xtreemfs-client_1.3.0rc1_amd64.deb ||
-  wget --directory-prefix=config/chroot_local-packages \
-  http://download.opensuse.org/repositories/home:/xtreemfs:/unstable/Debian_6.0/amd64/xtreemfs-client_1.3.0rc1_amd64.deb
+#test -f config/chroot_local-packages/xtreemfs-client_1.3.0.83_amd64.deb ||
+#  wget --directory-prefix=config/chroot_local-packages \
+#  http://download.opensuse.org/repositories/home:/xtreemfs:/unstable/Debian_6.0/amd64/xtreemfs-client_1.3.1.82_amd64.deb
 
 # we accept the terms of licence
-LICENCE_FILE="config/chroot_local-preseed/sun-licence"
-echo "sun-java6-bin shared/accepted-sun-dlj-v1-1 select true" >  $LICENCE_FILE
-echo "sun-java6-jre shared/accepted-sun-dlj-v1-1 select true" >> $LICENCE_FILE
+#LICENCE_FILE="config/chroot_local-preseed/sun-licence"
+#echo "sun-java6-bin shared/accepted-sun-dlj-v1-1 select true" >  $LICENCE_FILE
+#echo "sun-java6-jre shared/accepted-sun-dlj-v1-1 select true" >> $LICENCE_FILE
 
 mkdir -p config/chroot_local-includes/root/conf.cluster
 cp cust/conf.cluster/* config/chroot_local-includes/root/conf.cluster
@@ -40,12 +50,12 @@ install -D -m 755 cust/analyze_logs.pl config/chroot_local-includes/root/analyze
 install -D -m 644 cust/bydate.gnuplot config/chroot_local-includes/root/bydate.gnuplot
 install -D -m 644 cust/gmond.conf config/chroot_local-includes/etc/ganglia/gmond.conf
 
-mkdir -p config/chroot_local-includes/var/lib
-rm -rf config/chroot_local-includes/var/lib/sc-manager
+mkdir -p config/chroot_local-includes/usr/lib/scalaris/contrib
+rm -rf config/chroot_local-includes/usr/lib/scalaris/contrib/opennebula
 svn up ../webinterface-v2
-cp -R ../webinterface-v2 config/chroot_local-includes/var/lib/sc-manager
+cp -R ../webinterface-v2 config/chroot_local-includes/usr/lib/scalaris/contrib/opennebula
  
-install -D -m 755 cust/one.rb config/chroot_local-includes/var/lib/sc-manager/one.rb
+install -D -m 755 cust/one.rb config/chroot_local-includes/usr/lib/scalaris/contrib/opennebula/one.rb
 
 cp cust/post-install.sh config/chroot_local-hooks/00-post-install.sh
 #test -f /usr/share/live-helper/hooks/stripped && \
@@ -55,10 +65,12 @@ cp cust/post-install.sh config/chroot_local-hooks/00-post-install.sh
 if [ ! -d cust/pig ]; then
   mkdir -p config/chroot_local-includes/root
   svn checkout http://svn.apache.org/repos/asf/pig/trunk/ cust/pig
-else
-  svn update cust/pig
+#else
+#  svn update cust/pig
 fi
 
-cp -u cust/ContrailLogLoader.java cust/pig/contrib/piggybank/java/src/main/java/org/apache/pig/piggybank/storage/apachelog/
-pushd cust/pig && ant && cd contrib/piggybank/java && ant && popd
+if [ ! -f cust/pig/contrib/piggybank/java/piggybank.jar ]; then
+  cp -u cust/ContrailLogLoader.java cust/pig/contrib/piggybank/java/src/main/java/org/apache/pig/piggybank/storage/apachelog/
+  pushd cust/pig && ant && cd contrib/piggybank/java && ant && popd
+fi
 cp -u cust/pig/contrib/piggybank/java/piggybank.jar config/chroot_local-includes/root
